+++ workspace/tons-of-tests/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @creative-introvert/tons-of-tests
 
+## 1.1.1
+
+### Patch Changes
+
+- Feat: More subtle cli coloring.
+
+## 1.1.0
+
+### Minor Changes
+
+- Feat: Expose CLI.effect
+
 ## 1.0.0
 
 ### Major Changes